Move --run-rule test cases to runner as there is no SingleRuleRunner anymore
